Change project package name and add functionality for track, album and playlist download
This commit changes the project package name from "md_spotify_dl" to "main". The reason for this name change is to make the package name more conventional. This commit also adds new functionality for downloading Spotify tracks, albums and playlists. This includes the necessary command line interface setup using cobra for user interaction. A help function was also added for user guidance when invalid or no arguments are provided. The go.sum file was added to lock the dependencies to specific versions for reproducible builds.