Allow searching by track name along with Spotify URL
This commit allows a user to input either a Spotify URL or a track name when searching for music. If an incorrect URL or no URL is provided, instead of exiting the program, it assumes that the input is a track name and notifies the user about the search operation for that track. Also, adjusted the warning messages and the usage template to include the updated functionality. This update enhances the user flexibility by not mandating a Spotify URL.